Reuse ESS gRPC integration between Chariott and example code
