+++ common/Cargo.toml
@@ -5,14 +5,18 @@ edition = "2021"
 license = "MIT"
 
 [dependencies]
+async-trait = { workspace = true }
+ess = { path = "../ess" }
 prost = { workspace = true }
 prost-types = { workspace = true }
 regex = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 tokio = { workspace = true, features = ["signal"] }
 tokio-util = { workspace = true }
+tokio-stream = { workspace = true }
 tonic = { workspace = true }
 tracing = { workspace = true }
+uuid = { workspace = true }
 
 [build-dependencies]
 tonic-build = { workspace = true }
