+++ common/src/lib.rs
@@ -19,8 +19,11 @@
 /// Generic error handling
 pub mod error;
 
-// Configuration related utilites
+/// Configuration related utilites
 pub mod config;
+
+/// Integration of the event sub-system with the gRPC streaming contract.
+pub mod ess;
 
 /// Protobuf generated files.
 pub mod proto;
@@ -31,5 +34,5 @@ pub mod query;
 /// Graceful shutdown helpers
 pub mod shutdown;
 
-// Tokio utilities
+/// Tokio utilities
 pub mod tokio_runtime_fork;
