+++ examples/common/src/chariott/streaming.rs
@@ -2,48 +2,38 @@
 // Licensed under the MIT license.
 
 use async_trait::async_trait;
-use chariott_common::proto::common::{
-    fulfillment::Fulfillment, ReadFulfillment, ReadIntent, SubscribeFulfillment, SubscribeIntent,
+use chariott_common::{
+    ess::Ess as InnerEss,
+    proto::common::{
+        fulfillment::Fulfillment, ReadFulfillment, ReadIntent, SubscribeFulfillment,
+        SubscribeIntent,
+    },
 };
 use keyvalue::{InMemoryKeyValueStore, Observer};
-use std::{
-    fmt::Display,
-    hash::Hash,
-    sync::{Arc, RwLock},
-    time::SystemTime,
-};
-use tokio::spawn;
+use std::sync::RwLock;
 use tokio_stream::wrappers::ReceiverStream;
 use tonic::{Response, Status};
 
 use crate::chariott::proto::{
-    common::Value as ProtoValue,
+    common::value::Value as ProtoValue,
+    common::Value as ValueMessage,
     streaming::{channel_service_server::ChannelService, Event as ProtoEvent, OpenRequest},
 };
 
-const METADATA_KEY: &str = "x-chariott-channel-id";
+type EventId = Box<str>;
 
-type ClientId = Box<str>;
-type Ess<T> = ess::EventSubSystem<ClientId, EventId, (EventId, T), Result<ProtoEvent, Status>>;
+#[derive(Clone)]
+struct Ess<T>(InnerEss<(EventId, T)>);
 
-#[derive(Clone, Hash, Eq, PartialEq)]
-pub struct EventId(Box<str>);
-
-impl<T: Into<Box<str>>> From<T> for EventId {
-    fn from(value: T) -> Self {
-        Self(value.into())
-    }
-}
-
-impl Display for EventId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl<T: Clone> Observer<EventId, T> for Arc<Ess<T>> {
+impl<T: Clone + Send + 'static> Observer<EventId, T> for Ess<T> {
     fn on_set(&mut self, key: &EventId, value: &T) {
-        self.publish(key, (key.clone(), value.clone()));
+        self.0.as_ref().publish(key, (key.clone(), value.clone()));
+    }
+}
+
+impl<T> AsRef<InnerEss<(EventId, T)>> for Ess<T> {
+    fn as_ref(&self) -> &InnerEss<(EventId, T)> {
+        &self.0
     }
 }
 
@@ -52,19 +42,19 @@ impl<T: Clone> Observer<EventId, T> for Arc<Ess<T>> {
 /// value to be published, as long as that value can be transformed into a value
 /// which is compatible with the Proto contract.
 pub struct StreamingStore<T> {
-    ess: Arc<Ess<T>>,
-    store: RwLock<InMemoryKeyValueStore<EventId, T, Arc<Ess<T>>>>,
+    ess: Ess<T>,
+    store: RwLock<InMemoryKeyValueStore<EventId, T, Ess<T>>>,
 }
 
-impl<T: Clone> StreamingStore<T> {
+impl<T: Clone + Send + 'static> StreamingStore<T> {
     pub fn new() -> Self {
-        let ess = Arc::new(Ess::new());
-        let store = RwLock::new(InMemoryKeyValueStore::new(Some(Arc::clone(&ess))));
+        let ess = Ess(InnerEss::new());
+        let store = RwLock::new(InMemoryKeyValueStore::new(Some(ess.clone())));
         Self { ess, store }
     }
 }
 
-impl<T: Clone> Default for StreamingStore<T> {
+impl<T: Clone + Send + 'static> Default for StreamingStore<T> {
     fn default() -> Self {
         Self::new()
     }
@@ -81,23 +71,7 @@ where
         client_id: impl Into<Box<str>>,
         requested_subscriptions: impl IntoIterator<Item = EventId>,
     ) -> Result<(), Status> {
-        let subscriptions = self
-            .ess
-            .register_subscriptions(client_id.into(), requested_subscriptions)
-            .map_err(|_| Status::failed_precondition("Channel not open"))?;
-
-        for subscription in subscriptions {
-            spawn(subscription.serve(move |(key, value), seq| {
-                Ok(ProtoEvent {
-                    source: key.to_string(),
-                    value: Some(value.into()),
-                    seq,
-                    timestamp: Some(SystemTime::now().into()),
-                })
-            }));
-        }
-
-        Ok(())
+        self.ess.as_ref().serve_subscriptions(client_id, requested_subscriptions, |(_, v)| v.into())
     }
 
     /// Read a value from the store.
@@ -120,13 +94,9 @@ where
 
     async fn open(
         &self,
-        _: tonic::Request<OpenRequest>,
+        request: tonic::Request<OpenRequest>,
     ) -> Result<Response<Self::OpenStream>, Status> {
-        let id: ClientId = uuid::Uuid::new_v4().to_string().into();
-        let (_, receiver_stream) = self.ess.read_events(id.clone());
-        let mut response = Response::new(receiver_stream);
-        response.metadata_mut().insert(METADATA_KEY, id.to_string().try_into().unwrap());
-        Ok(response)
+        self.ess.as_ref().open(request).await
     }
 }
 
@@ -150,6 +120,8 @@ where
 
     fn read(&self, intent: ReadIntent) -> Fulfillment {
         let value = self.get(&intent.key.into());
-        Fulfillment::Read(ReadFulfillment { value: value.map(|v| v.into()) })
+        Fulfillment::Read(ReadFulfillment {
+            value: Some(ValueMessage { value: value.map(|v| v.into()) }),
+        })
     }
 }
